Delete the old core data stack and related files
