+++ Tests/Source/ManagedObjectContext/DatabaseBaseTest.swift
@@ -34,10 +34,7 @@ import WireTesting
     override public func setUp() {
         super.setUp()
         self.clearStorageFolder()
-        MainPersistentStoreRelocator.hostBundleIdentifier = Bundle.main.bundleIdentifier!
         try! FileManager.default.createDirectory(at: self.applicationContainer, withIntermediateDirectories: true)
-        let legacyDatabaseDirectory = applicationContainer.appendingPathComponent(Bundle.main.bundleIdentifier!)
-        try! FileManager.default.createDirectory(at: legacyDatabaseDirectory, withIntermediateDirectories: true)
     }
     
     override public func tearDown() {
@@ -58,25 +55,6 @@ import WireTesting
         }
 
         return stack
-    }
-
-    /// Create storage stack at a legacy location
-    @objc public func createLegacyStore(searchPath: FileManager.SearchPathDirectory) {
-        let directory = FileManager.default.urls(for: searchPath, in: .userDomainMask).first!
-        self.createLegacyStore(filePath: directory.appendingStoreFile())
-    }
-    
-    /// Create storage stack at a legacy location
-    @objc public func createLegacyStore(filePath: URL,
-                                        customization: ((ManagedObjectContextDirectory)->())? = nil) {
-
-        guard let persistentStoreCoordinator = try? NSPersistentStoreCoordinator.create(
-                storeFile: filePath,
-                applicationContainer: applicationContainer) else {
-            return
-        }
-        
-        self.createDummyExternalSupportFileForDatabase(storeFile: filePath)
     }
     
     /// Create a session in the keystore directory for the given account
@@ -105,10 +83,6 @@ import WireTesting
     public func clearStorageFolder() {
         let url = self.applicationContainer
         try? FileManager.default.removeItem(at: url)
-        
-        self.previousDatabaseLocations.forEach {
-            try? FileManager.default.removeItem(at: $0)
-        }
     }
     
     /// Creates some dummy Core Data store support file
@@ -119,33 +93,4 @@ import WireTesting
         try! self.mediumJPEGData().write(to: supportPath.appendingPathComponent("image.dat"))
     }
     
-    /// Extensions after the database file name
-    /// This is needed to expose Swift-only property to Obj-c
-    public static var databaseFileExtensions: [String] {
-        return PersistentStoreRelocator.storeFileExtensions
-    }
-    
-    /// Previous locations where the database was stored
-    var previousDatabaseLocations: [URL] {
-        return [
-            FileManager.default.urls(for: .cachesDirectory, in: .userDomainMask).first!,
-            FileManager.default.urls(for: .applicationSupportDirectory, in: .userDomainMask).first!,
-            self.applicationContainer.appendingPathComponent(Bundle.main.bundleIdentifier!),
-            self.applicationContainer.appendingPathComponent(Bundle.main.bundleIdentifier!).appendingPathComponent(accountID.uuidString).appendingPathComponent("store")
-        ]
-    }
-    
-    var previousDatabaseLocationsBeforeMultiAccountSupport: ArraySlice<URL> {
-        return previousDatabaseLocations.prefix(3)
-    }
-
-    /// Previous locations where the keystore was stored
-    var previousKeyStoreLocations: [URL] {
-        return [
-            FileManager.default.urls(for: .cachesDirectory, in: .userDomainMask).first!,
-            FileManager.default.urls(for: .applicationSupportDirectory, in: .userDomainMask).first!,
-            self.applicationContainer,
-            self.applicationContainer.appendingPathComponent(Bundle.main.bundleIdentifier!).appendingPathComponent(accountID.uuidString)
-        ]
-    }
 }
