+++ Tests/Source/Model/UserClient/UserClientKeyStoreTests.swift
@@ -46,7 +46,7 @@ class UserClientKeysStoreTests: OtrBaseTest {
     
     func cleanOTRFolder() {
         let fm = FileManager.default
-        var paths = UserClientKeysStore.possibleLegacyKeyStores(applicationContainer: OtrBaseTest.sharedContainerURL, accountIdentifier: accountID).map{$0.path}
+        var paths: [String] = []
         if let accountID = accountID {
             paths.append(OtrBaseTest.otrDirectoryURL(accountIdentifier: accountID).path)
         }
@@ -129,57 +129,4 @@ class UserClientKeysStoreTests: OtrBaseTest {
         return folder
     }
 
-    func testThatItMovesTheOTRFolderToTheGivenURL() {
-        
-        let possibleLegacyKeyStores = UserClientKeysStore.possibleLegacyKeyStores(applicationContainer: OtrBaseTest.sharedContainerURL, accountIdentifier: accountID)
-        
-        for legacyKeyStoreLocation in possibleLegacyKeyStores {
-            // given
-            self.sut = nil
-            self.cleanOTRFolder()
-            let accountFolder = CoreDataStack.accountDataFolder(accountIdentifier: self.accountID, applicationContainer: OtrBaseTest.sharedContainerURL)
-            let data = "foo".data(using: String.Encoding.utf8)!
-            _ = self.createLegacyOTRFolderWithDummyFile(fileName: "dummy.txt", data: data, folder: legacyKeyStoreLocation)
-            
-            // when
-            UserClientKeysStore.migrateIfNeeded(accountIdentifier: accountID, accountDirectory: accountFolder, applicationContainer: OtrBaseTest.sharedContainerURL)
-            
-            // then
-            let expectedFileURL = FileManager.keyStoreURL(accountDirectory: accountFolder, createParentIfNeeded: false).appendingPathComponent("dummy.txt")
-            let fooData = try! Data(contentsOf: expectedFileURL)
-            let fooString = String(data: fooData, encoding: String.Encoding.utf8)!
-            XCTAssertEqual(fooString, "foo")
-            XCTAssertFalse(UserClientKeysStore.needToMigrateIdentity(applicationContainer: OtrBaseTest.sharedContainerURL, accountIdentifier: accountID))
-        }
-    }
-    
-    func testThatItMovesTheOTRFolderAgainIfTheFirstMigrationFailed() {
-        
-        // this is testing migrating half-way (not completely copying the OTR folder)
-        // and then restarting migration after a crash
-        
-        // given
-        self.sut = nil
-        self.cleanOTRFolder()
-        let fileName = "dummy.txt"
-        let data = "foo".data(using: String.Encoding.utf8)!
-        let accountFolder = CoreDataStack.accountDataFolder(accountIdentifier: self.accountID, applicationContainer: OtrBaseTest.sharedContainerURL)
-        
-        // first migration
-        _ = self.createLegacyOTRFolderWithDummyFile(fileName: "whatever.dat", data: data)
-        UserClientKeysStore.migrateIfNeeded(accountIdentifier: accountID, accountDirectory: accountFolder, applicationContainer: OtrBaseTest.sharedContainerURL)
-        
-        // to pretend it's not done, re-create the folder
-        _ = self.createLegacyOTRFolderWithDummyFile(fileName: fileName, data: data)
-        
-        // when
-        UserClientKeysStore.migrateIfNeeded(accountIdentifier: accountID, accountDirectory: accountFolder, applicationContainer: OtrBaseTest.sharedContainerURL)
-        
-        // then
-        let fooData = try! Data(contentsOf: OtrBaseTest.otrDirectoryURL(accountIdentifier: accountID).appendingPathComponent(fileName))
-        let fooString = String(data: fooData, encoding: String.Encoding.utf8)!
-        XCTAssertEqual(fooString, "foo")
-        XCTAssertFalse(UserClientKeysStore.needToMigrateIdentity(applicationContainer: OtrBaseTest.sharedContainerURL, accountIdentifier: accountID))
-        
-    }
 }
